refactor: extract save operations + fix imports
